intern-project/paras/project-DHAP-42/requirements.txt
# Airflow
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.6.1

# Data Processing
pandas==2.1.3
pyarrow==14.0.1
numpy==1.25.2

# MinIO
minio==7.2.0
boto3==1.34.0
botocore==1.34.0

# Schema Validation
pydantic==2.5.0
PyYAML==6.0.1
jsonschema==4.20.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==1.4.50

# Utilities
python-dotenv==1.0.0
requests==2.31.0
